Transferred citation reference functions from metadata to their own module
