+++ metadata_processing.py
@@ -50,9 +50,6 @@ class CitingCasesError(TypedDict):
     error: str
     metadata: None
 
-#CitingCasesResult = Union[CitingCasesSuccess, CitingCasesError]
-CaseReferenceType = Literal["citedCases", "citingCases", "citedLegislations"]
-
 def _parse_citation(citation: str) -> Optional[Dict[str, Any]]:
     """
     Parse a citation using the legal citation parser.
@@ -73,7 +70,6 @@ def _parse_citation(citation: str) -> Optional[Dict[str, Any]]:
 
 def get_metadata_from_citation(
     citation: str,
-    include_relations: bool = True,
 ) -> Optional[CitationMetadata]:
     """Extract all metadata from a citation using the legal citation parser."""
     try:
@@ -116,78 +112,9 @@ def get_metadata_from_citation(
             "keywords": citation_data.get('keywords', []),
             "categories": citation_data.get('categories', []),
         }
-
-        if include_relations:
-            cited_cases_result = get_case_relations(citation, "citedCases")
-            citing_cases_result = get_case_relations(citation, "citingCases")
-            metadata["cited_cases"] = cited_cases_result.get("citedCases", [])
-            metadata["citing_cases"] = citing_cases_result.get("citingCases", [])
         
         return metadata
         
     except Exception as e:
         logger.error("Error getting metadata from citation: %s", e)
         return None
-
-def get_case_relations(
-    citation: str,
-    caseref_type: CaseReferenceType = "citingCases",
-    timeout_seconds: int = REQUEST_TIMEOUT_SECONDS,
-) -> Dict[str, Any]:
-    """
-    Fetch related cases/legislation from the CanLII case citator API and
-    add them into the metadata dict under the given caseref_type key.
-
-    caseref_type: "citedCases", "citingCases", or "citedLegislations"
-    """
-    try:
-        api_key = Config.CANLII_API_KEY
-        if not api_key:
-            raise ValueError("CANLII_API_KEY not found in environment variables")
-        
-        citation_data = _parse_citation(citation)
-        if not citation_data or 'court' not in citation_data or 'uid' not in citation_data:
-            raise ValueError(f"Could not parse citation: {citation}")
-            
-        # Convert court name to lowercase and clean up decision code
-        court_name = citation_data['court'].lower()
-        decision_code = citation_data['uid'].lower().replace('-', '')
-
-        # CaseReferenceType is encoded in the URL path per CanLII docs
-        url = (
-            f"https://api.canlii.org/v1/caseCitator/en/"
-            f"{court_name}/{decision_code}/{caseref_type}"
-            f"?api_key={api_key}"
-        )
-        logger.info("Making CanLII API request")  # Don't log the URL with API key
-
-        response = requests.get(url, timeout=timeout_seconds)
-
-        if response.status_code == 429:
-            logger.warning("CanLII API rate limit reached")
-            return {
-                "error": "Rate limit reached",
-                caseref_type: [],
-                "metadata": citation_data.get("metadata", {}),
-            }
-            
-        if response.status_code != 200:
-            logger.error(f"CanLII API error: {response.status_code}")
-            return {
-                "error": f"API error: {response.status_code}",
-                caseref_type: [],
-                "metadata": citation_data.get("metadata", {}),
-            }
-
-        data = response.json()
-        items = data.get(caseref_type, [])
-
-        return {
-            caseref_type: items,
-            "error": None,
-            "metadata": citation_data.get("metadata", {}) or {},
-        }
-
-    except Exception as e:
-        logger.error("Error fetching case relations: %s", e)
-        return {"error": str(e), "citingCases": [], "citedCases": [], "metadata": None}
